requirements.txt
# DigiComply Cyber Shield — Python Dependencies

# Core Framework
fastapi==0.115.0
uvicorn==0.30.0
pydantic>=2.0

# Scanning
dnspython==2.6.1
requests>=2.31.0
python-nmap==0.7.1

# AI Layer (policy generation, report writing, alert triage)
anthropic>=0.39.0
python-dotenv>=1.0

# PDF Report Generation
reportlab>=4.0
jinja2>=3.1.0

# Form parsing
python-multipart>=0.0.6

# Phishing (GoPhish API client)
# GoPhish itself runs as Docker container

# Email Alerts
# sendgrid>=6.0  # or use smtp directly

# Database (choose one)
# supabase>=2.0  # Supabase (free tier)
# sqlalchemy>=2.0  # Direct PostgreSQL

# Scheduling
# celery>=5.3  # For cron jobs (weekly scans, monthly reports)
# redis>=5.0   # Celery broker

# Optional: Nuclei vulnerability scanning (Go binary, not pip)
# Install separately: go install github.com/projectdiscovery/nuclei/v3/cmd/nuclei@latest
